rbac_backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-bootstrap-form==3.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.2
tzdata==2024.2
